Reduce extra work in filter.
After the first all, we have the values, we do not once again "resolve" them. We merely must resolve the results of the filterFn with another All.

At first glance this seems pointless, but unfortunately, no good test exist to short-circuit the all for an object value, as an object may be a thenable, but a true thenable check is deopting, and can only happen once. Interestingly though, all can be optimized for values, such as booleans which is a likely result of the filterFn.

Ultimately this reduces unneeded work, resulting in better performance. This is especially so until Promise.all is properly optimized, but even then this will be the faster path.