[fixes #330] “works” in Nashorn
NOTE: as Nashorn has real threads, it does not have the same
run-to-completion guarentee as the JavaScript you may be familiar with.
As such if you are not extremely careful in how you use these threads
you will easily put browser/node JavaScript libraries in a bad state.
RSVP included.

It would be fun to fork RSVP and ensure it’s thread safety in such
environments, but as no other environment has these requirements it not
likely not happen any time soon.

* also fix `RSVP.on(‘error’,` on vertex

Also, the test suite itself does not yet run in Nashorn so only some
manual testing “proved” it would work. I have relatively good confidence
in this, but at some point soon it would be better to get the test suite
running on Nashorn. I do suspect it will likely be a sizable investment.