Make example code import and use the right modules in the right way
The former example where not technically wrong but pretty misleading,
since they suggested `Promise` (the one that can be used in
`new Promise`) is the default export of the `rsvp` package when it isn't.
Ideally when using the default import it should be called `RSVP` rather
then `Promise` but in fact neither of the code example require to use
that default export at all so I removed using that in the examples
entirely.